Casted types in the arc method

+++ test-modules/movement/main.cpp
@@ -82,8 +82,8 @@ void rotateRight() {
 // 0 is max forward
 // 1 is max right
 void arc(double direction) {
-    setLeft(5 * direction);
-    setRight(-5 * direction);
+    setLeft((int) 5 * direction);
+    setRight((int) -5 * direction);
 }
 
 // ******************** TESTING ONLY ******************** //
